adjust frontend search request format
