+++ frontend/src/app/models/elasticsearch.ts
@@ -68,3 +68,20 @@ export interface SearchHit {
     _source: FieldValues;
     highlight?: HighlightResult;
 }
+
+export type EsQuerySorted = EsQuery & {
+    sort: { [fieldName: string]: 'desc' | 'asc' }[];
+};
+
+export interface EsQuery {
+    aborted?: boolean;
+    completed?: Date;
+    query: EsSearchClause | BooleanQuery | EsFilter;
+    highlight?: unknown;
+    transferred?: number;
+}
+
+export interface Aggregator {
+    name: string;
+    size: number;
+}
