+++ frontend/src/app/models/query.ts
@@ -2,7 +2,7 @@ import { convertToParamMap, ParamMap } from '@angular/router';
 import * as _ from 'lodash';
 import { combineLatest, Subject } from 'rxjs';
 import { Corpus, CorpusField, EsFilter, SortBy, SortConfiguration, SortDirection, } from '../models/index';
-import { EsQuery } from '../services';
+import { EsQuery } from '../models';
 import { combineSearchClauseAndFilters, makeHighlightSpecification } from '../utils/es-query';
 import {
     filtersFromParams, highlightFromParams, omitNullParameters, queryFiltersToParams,
