+++ frontend/src/app/models/search-requests.ts
@@ -1,0 +1,6 @@
+import { EsQuery } from './elasticsearch';
+
+export interface QueryParameters {
+    es_query: EsQuery;
+    tags?: number[];
+}
